Removed defines to promote rapid production.
Defines are great in theory, but in most scenarios we'd want variables
representing the terrain information to provide flexibility in our
framework. Since this is a prototype, the numbs won't change, but tweaking
them separately from everything else related to terrain is a pain.
Especially since they're only ever going to be referenced once as initial
values.