Tweaked build paths a bit more...
To clarify, I discovered that the ProjectDir variable contains about the
following:

C:\projects\Nimbus\

This means that a path such as $(ProjectDir)\build\etc completes to:

C:\projects\Nimbus\\build\etc

Note the double backslashes. We discussed their relevance a few times
before, but I'm wondering if removing our own backslash will fix the
issues we've had. If not, we'll switch back to the previous solution.